config.properties and it's implementation to make Files-Upload directory delete or not during testing.

+++ src/main/java/spring/framework/books/startup/OnStartup.java
@@ -3,33 +3,50 @@ package spring.framework.books.startup;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 import org.springframework.context.ApplicationListener;
+import org.springframework.context.EnvironmentAware;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.context.event.ContextRefreshedEvent;
+import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Component;
 
 import javax.transaction.Transactional;
 import java.io.File;
 import java.io.IOException;
+import java.util.Objects;
 
 
 @Slf4j
 @Component
-public class OnStartup implements ApplicationListener<ContextRefreshedEvent> {
+@PropertySource("classpath:config.properties")
+public class OnStartup implements ApplicationListener<ContextRefreshedEvent>, EnvironmentAware {
+
+    private static Environment environment;
+
+    @Override
+    public void setEnvironment(Environment environment) {
+        OnStartup.environment = environment;
+    }
 
 
     @Transactional
     @Override
     public void onApplicationEvent(ContextRefreshedEvent event) {
 
-         File file=new File("../../../Files-Upload");
+        if (Objects.requireNonNull(environment.getProperty("IS_TESTING_ENABLED")).equalsIgnoreCase("1")){
+            File file=new File("Files-Upload");
 
-        try {
-            System.out.println("Deleting Directory!!");
-            FileUtils.deleteDirectory(file);
-            System.out.println("Directory Deleted!!");
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+            try {
+                System.out.println("Deleting Directory!!");
+                FileUtils.deleteDirectory(file);
+                System.out.println("Directory Deleted!!");
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
         }
 
 
+
     }
+
+
 }
